+++ layabase/mongo.py
@@ -578,40 +578,46 @@ class Column:
         elif isinstance(value, list) and self.field_type != list:
             if value:  # Discard empty list as filter on non list field
                 mongo_values = [
-                    self._deserialize_value(value_in_list) for value_in_list in value
+                    self._deserialize_value(value_in_list)
+                    if value_in_list is not None
+                    else None
+                    for value_in_list in value
                 ]
-                if self.get_default_value(filters) in mongo_values:
-                    or_filter = filters.setdefault("$or", [])
-                    or_filter.append({self.name: {"$exists": False}})
-                    or_filter.append({self.name: {"$in": mongo_values}})
-                else:
-                    mongo_filters = {}
-                    other_values = []
-                    for val in mongo_values:
-                        if isinstance(val, tuple):
-                            mongo_filters.update({_operators[val[0]]: val[1]})
-                        else:
-                            other_values.append(val)
-
-                    if mongo_filters and other_values:
-                        or_filter = filters.setdefault("$or", [])
-                        or_filter.append({self.name: mongo_filters})
-                        or_filter.append({self.name: {"$in": mongo_values}})
-                    else:
-                        filters[self.name] = (
-                            {"$in": other_values} if other_values else mongo_filters
+                comparison_filters = {}
+                equality_values = []
+                for mongo_value in mongo_values:
+                    if isinstance(mongo_value, tuple):
+                        comparison_filters.update(
+                            {_operators[mongo_value[0]]: mongo_value[1]}
                         )
+                    else:
+                        equality_values.append(mongo_value)
+
+                filters_on_field = []
+
+                if self.get_default_value(filters) in equality_values:
+                    filters_on_field.append({self.name: {"$exists": False}})
+
+                if equality_values:
+                    filters_on_field.append({self.name: {"$in": equality_values}})
+
+                if comparison_filters:
+                    filters_on_field.append({self.name: comparison_filters})
+
+                if len(filters_on_field) == 1:
+                    filters[self.name] = filters_on_field[0][self.name]
+                else:
+                    filters.setdefault("$or", []).extend(filters_on_field)
         else:
             mongo_value = self._deserialize_value(value)
             if self.get_default_value(filters) == mongo_value:
                 or_filter = filters.setdefault("$or", [])
                 or_filter.append({self.name: {"$exists": False}})
                 or_filter.append({self.name: mongo_value})
+            elif isinstance(mongo_value, tuple):
+                filters[self.name] = {_operators[mongo_value[0]]: mongo_value[1]}
             else:
-                if isinstance(mongo_value, tuple):
-                    filters[self.name] = {_operators[mongo_value[0]]: mongo_value[1]}
-                else:
-                    filters[self.name] = mongo_value
+                filters[self.name] = mongo_value
 
     def deserialize_insert(self, document: dict):
         """
@@ -673,13 +679,7 @@ class Column:
         :param value: Received field value.
         :return Mongo valid value.
         """
-        if value is None:
-            return None
-
-        if isinstance(value, str):
-            value = iso8601.parse_date(value)
-
-        return value
+        return iso8601.parse_date(value) if isinstance(value, str) else value
 
     def _deserialize_date(self, value):
         """
@@ -687,9 +687,6 @@ class Column:
         :param value: Received field value.
         :return Mongo valid value.
         """
-        if value is None:
-            return None
-
         if isinstance(value, str):
             value = iso8601.parse_date(value)
         elif isinstance(value, datetime.date):
@@ -711,9 +708,6 @@ class Column:
         :param value: Received field value.
         :return Mongo valid value.
         """
-        if value is None:
-            return None
-
         # Enum cannot be stored in Mongo, use enum value instead
         if isinstance(value, enum.Enum):
             value = value.value
@@ -729,13 +723,7 @@ class Column:
         :param value: Received field value.
         :return Mongo valid value.
         """
-        if value is None:
-            return None
-
-        if not isinstance(value, ObjectId):
-            value = ObjectId(value)
-
-        return value
+        return value if isinstance(value, ObjectId) else ObjectId(value)
 
     def _deserialize_int(self, value):
         """
@@ -744,13 +732,7 @@ class Column:
         :param value: Received field value.
         :return Mongo valid value.
         """
-        if value is None:
-            return None
-
-        if isinstance(value, str):
-            value = int(value)
-
-        return value
+        return int(value) if isinstance(value, str) else value
 
     def _deserialize_float(self, value):
         """
@@ -759,13 +741,7 @@ class Column:
         :param value: Received field value.
         :return Mongo valid value.
         """
-        if value is None:
-            return None
-
-        if isinstance(value, str):
-            value = float(value)
-
-        return value
+        return float(value) if isinstance(value, str) else value
 
     def _deserialize_str(self, value):
         """
@@ -774,13 +750,7 @@ class Column:
         :param value: Received field value.
         :return Mongo valid value.
         """
-        if value is None:
-            return None
-
-        if not isinstance(value, str):
-            value = str(value)
-
-        return value
+        return value if isinstance(value, str) else str(value)
 
     def serialize(self, document: dict):
         """
@@ -914,20 +884,11 @@ class DictColumn(Column):
             raise Exception("fields or get_fields must be provided.")
 
         self._default_fields = fields or {}
-        self._get_fields = (
-            get_fields if get_fields else lambda model_as_dict: self._default_fields
-        )
+        self._get_fields = get_fields or (lambda model_as_dict: self._default_fields)
 
-        if index_fields:
-            self._get_all_index_fields = (
-                get_index_fields
-                if get_index_fields
-                else lambda model_as_dict: index_fields
-            )
-            self._default_index_fields = index_fields
-        else:
-            self._get_all_index_fields = self._get_fields
-            self._default_index_fields = self._default_fields
+        self._get_all_index_fields = get_index_fields or (
+            (lambda model_as_dict: index_fields) if index_fields else self._get_fields
+        )
 
         if bool(
             kwargs.get("is_nullable", True)
@@ -965,18 +926,6 @@ class DictColumn(Column):
             self._get_fields(model_as_dict),
         )
 
-    def _default_index_description_model(self):
-        """
-        :return: A CRUDModel describing every index fields.
-        """
-        from layabase._database_mongo import _CRUDModel
-
-        return type(
-            f"{self.name}_DefaultIndexDescriptionModel",
-            (_CRUDModel,),
-            self._default_index_fields,
-        )
-
     def _index_description_model(self, model_as_dict: dict):
         """
         :param model_as_dict: Data provided by the user.
@@ -993,10 +942,6 @@ class DictColumn(Column):
     def _get_index_fields(
         self, index_type: IndexType, model_as_dict: Union[dict, None], prefix: str
     ) -> List[str]:
-        if model_as_dict is None:
-            return self._default_index_description_model()._get_index_fields(
-                index_type, None, f"{prefix}{self.name}."
-            )
         return self._index_description_model(model_as_dict)._get_index_fields(
             index_type, model_as_dict, f"{prefix}{self.name}."
         )
@@ -1174,7 +1119,7 @@ class ListColumn(Column):
     def validate_update(self, document: dict) -> dict:
         errors = Column.validate_update(self, document)
         if not errors:
-            values = document[self.name]
+            values = document[self.name] or []
             for index, value in enumerate(values):
                 document_with_list_item = {**document, self.name: value}
                 list_item_errors = self.list_item_column.validate_update(
@@ -1191,8 +1136,9 @@ class ListColumn(Column):
     def deserialize_update(self, document: dict):
         values = document.get(self.name)
         if values is None:
-            # Ensure that None value are not stored to save space and allow to change default value.
-            document.pop(self.name, None)
+            if not self._store_none:
+                # Ensure that None value are not stored to save space and allow to change default value.
+                document.pop(self.name, None)
         else:
             new_values = []
             for value in values: