+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.3] - 2019-10-10
+### Fixed
+- Update [SQLAlchemy](https://docs.sqlalchemy.org/en/13/changelog/index.html) to version 1.3.10
+
+## [3.0.2] - 2019-10-10
+### Fixed
+- [Mongo] Allow to update a ListColumn field to None.
+- [Mongo] Allow to provide indexes for a DictColumn via get_index_fields function only.
+- [Mongo] Allow to query by comparison signs AND default value AND equality at the same time
+
+### Removed
+- [Mongo] Do not expect BulkWriteError on update many as it should be triggered only on insertion.
+
 ## [3.0.1] - 2019-10-08
 ### Fixed
 - The ability to link a Mongo controller to a collection after loading the database was lost in 3.0.0. It is now available via layabase.mongo.link(controller, base)
@@ -113,7 +126,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Initial release.
 
-[Unreleased]:https://github.tools.digital.engie.com/gempy/layabase/compare/v3.0.1...HEAD
+[Unreleased]:https://github.tools.digital.engie.com/gempy/layabase/compare/v3.0.3...HEAD
+[3.0.3]:https://github.tools.digital.engie.com/gempy/layabase/compare/v3.0.2...v3.0.3
+[3.0.2]:https://github.tools.digital.engie.com/gempy/layabase/compare/v3.0.1...v3.0.2
 [3.0.1]:https://github.tools.digital.engie.com/gempy/layabase/compare/v3.0.0...v3.0.1
 [3.0.0]:https://github.tools.digital.engie.com/gempy/layabase/compare/v2.0.1...v3.0.0
 [2.0.1]:https://github.tools.digital.engie.com/gempy/layabase/compare/v2.0.0...v2.0.1